# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for sito annunci
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Annuncio {
  id: number;
  titolo: string;
  descrizione: string;
  /** @nullable */
  prezzo?: number | null;
  categoria: string;
  citta: string;
  contatto: string;
  /** @nullable */
  immagineUrl?: string | null;
  inEvidenza: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AnnunciListResponse {
  annunci: Annuncio[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateAnnuncioBody {
  titolo: string;
  descrizione: string;
  /** @nullable */
  prezzo?: number | null;
  categoria: string;
  citta: string;
  contatto: string;
  /** @nullable */
  immagineUrl?: string | null;
}

export interface UpdateAnnuncioBody {
  titolo?: string;
  descrizione?: string;
  /** @nullable */
  prezzo?: number | null;
  categoria?: string;
  citta?: string;
  contatto?: string;
  /** @nullable */
  immagineUrl?: string | null;
}

export interface Categoria {
  id: string;
  nome: string;
  icona: string;
  descrizione: string;
  count: number;
}

export interface CittaCount {
  citta: string;
  count: number;
}

export interface CategoriaCount {
  categoria: string;
  count: number;
}

export interface SiteStats {
  totaleAnnunci: number;
  annunciOggi: number;
  totaleCategorie: number;
  cittaPiuAttive: CittaCount[];
  annunciPerCategoria: CategoriaCount[];
}

export interface ForumThread {
  id: number;
  titolo: string;
  corpo: string;
  categoria: string;
  autore: string;
  /** @nullable */
  autoreClerkId?: string | null;
  risposteCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ForumReply {
  id: number;
  threadId: number;
  corpo: string;
  autore: string;
  /** @nullable */
  autoreClerkId?: string | null;
  createdAt: string;
}

export interface ForumThreadWithReplies {
  thread: ForumThread;
  replies: ForumReply[];
}

export interface ForumListResponse {
  threads: ForumThread[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateForumThreadBody {
  titolo: string;
  corpo: string;
  categoria: string;
  autore: string;
  /** @nullable */
  autoreClerkId?: string | null;
}

export interface CreateForumReplyBody {
  corpo: string;
  autore: string;
  /** @nullable */
  autoreClerkId?: string | null;
}

export type ListAnnunciParams = {
  /**
   * @nullable
   */
  categoria?: string | null;
  /**
   * @nullable
   */
  q?: string | null;
  /**
   * @nullable
   */
  prezzoMin?: number | null;
  /**
   * @nullable
   */
  prezzoMax?: number | null;
  /**
   * @nullable
   */
  citta?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type ListForumThreadsParams = {
  /**
   * @nullable
   */
  categoria?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};