# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for sito annunci
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Lista annunci con filtri
 */
export const ListAnnunciQueryParams = zod.object({
  categoria: zod.coerce.string().nullish(),
  q: zod.coerce.string().nullish(),
  prezzoMin: zod.coerce.number().nullish(),
  prezzoMax: zod.coerce.number().nullish(),
  citta: zod.coerce.string().nullish(),
  page: zod.coerce.number().nullish(),
  limit: zod.coerce.number().nullish(),
});

export const ListAnnunciResponse = zod.object({
  annunci: zod.array(
    zod.object({
      id: zod.number(),
      titolo: zod.string(),
      descrizione: zod.string(),
      prezzo: zod.number().nullish(),
      categoria: zod.string(),
      citta: zod.string(),
      contatto: zod.string(),
      immagineUrl: zod.string().nullish(),
      inEvidenza: zod.boolean(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Crea un nuovo annuncio
 */
export const CreateAnnuncioBody = zod.object({
  titolo: zod.string(),
  descrizione: zod.string(),
  prezzo: zod.number().nullish(),
  categoria: zod.string(),
  citta: zod.string(),
  contatto: zod.string(),
  immagineUrl: zod.string().nullish(),
});

/**
 * @summary Annunci pubblicati di recente
 */
export const GetAnnunciRecentiResponseItem = zod.object({
  id: zod.number(),
  titolo: zod.string(),
  descrizione: zod.string(),
  prezzo: zod.number().nullish(),
  categoria: zod.string(),
  citta: zod.string(),
  contatto: zod.string(),
  immagineUrl: zod.string().nullish(),
  inEvidenza: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetAnnunciRecentiResponse = zod.array(
  GetAnnunciRecentiResponseItem,
);

/**
 * @summary Annunci in evidenza/sponsorizzati
 */
export const GetAnnunciInEvidenzaResponseItem = zod.object({
  id: zod.number(),
  titolo: zod.string(),
  descrizione: zod.string(),
  prezzo: zod.number().nullish(),
  categoria: zod.string(),
  citta: zod.string(),
  contatto: zod.string(),
  immagineUrl: zod.string().nullish(),
  inEvidenza: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetAnnunciInEvidenzaResponse = zod.array(
  GetAnnunciInEvidenzaResponseItem,
);

/**
 * @summary Dettaglio annuncio
 */
export const GetAnnuncioParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnnuncioResponse = zod.object({
  id: zod.number(),
  titolo: zod.string(),
  descrizione: zod.string(),
  prezzo: zod.number().nullish(),
  categoria: zod.string(),
  citta: zod.string(),
  contatto: zod.string(),
  immagineUrl: zod.string().nullish(),
  inEvidenza: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Aggiorna annuncio
 */
export const UpdateAnnuncioParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAnnuncioBody = zod.object({
  titolo: zod.string().optional(),
  descrizione: zod.string().optional(),
  prezzo: zod.number().nullish(),
  categoria: zod.string().optional(),
  citta: zod.string().optional(),
  contatto: zod.string().optional(),
  immagineUrl: zod.string().nullish(),
});

export const UpdateAnnuncioResponse = zod.object({
  id: zod.number(),
  titolo: zod.string(),
  descrizione: zod.string(),
  prezzo: zod.number().nullish(),
  categoria: zod.string(),
  citta: zod.string(),
  contatto: zod.string(),
  immagineUrl: zod.string().nullish(),
  inEvidenza: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Elimina annuncio
 */
export const DeleteAnnuncioParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Lista categorie disponibili
 */
export const ListCategorieResponseItem = zod.object({
  id: zod.string(),
  nome: zod.string(),
  icona: zod.string(),
  descrizione: zod.string(),
  count: zod.number(),
});
export const ListCategorieResponse = zod.array(ListCategorieResponseItem);

/**
 * @summary Statistiche del sito (totale annunci, per categoria, ecc.)
 */
export const GetStatsResponse = zod.object({
  totaleAnnunci: zod.number(),
  annunciOggi: zod.number(),
  totaleCategorie: zod.number(),
  cittaPiuAttive: zod.array(
    zod.object({
      citta: zod.string(),
      count: zod.number(),
    }),
  ),
  annunciPerCategoria: zod.array(
    zod.object({
      categoria: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Lista categorie del forum
 */
export const GetForumCategorieResponseItem = zod.string();
export const GetForumCategorieResponse = zod.array(
  GetForumCategorieResponseItem,
);

/**
 * @summary Lista discussioni forum
 */
export const ListForumThreadsQueryParams = zod.object({
  categoria: zod.coerce.string().nullish(),
  page: zod.coerce.number().nullish(),
  limit: zod.coerce.number().nullish(),
});

export const ListForumThreadsResponse = zod.object({
  threads: zod.array(
    zod.object({
      id: zod.number(),
      titolo: zod.string(),
      corpo: zod.string(),
      categoria: zod.string(),
      autore: zod.string(),
      autoreClerkId: zod.string().nullish(),
      risposteCount: zod.number(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Crea nuova discussione
 */
export const CreateForumThreadBody = zod.object({
  titolo: zod.string(),
  corpo: zod.string(),
  categoria: zod.string(),
  autore: zod.string(),
  autoreClerkId: zod.string().nullish(),
});

/**
 * @summary Dettaglio discussione con risposte
 */
export const GetForumThreadParams = zod.object({
  id: zod.coerce.number(),
});

export const GetForumThreadResponse = zod.object({
  thread: zod.object({
    id: zod.number(),
    titolo: zod.string(),
    corpo: zod.string(),
    categoria: zod.string(),
    autore: zod.string(),
    autoreClerkId: zod.string().nullish(),
    risposteCount: zod.number(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
  replies: zod.array(
    zod.object({
      id: zod.number(),
      threadId: zod.number(),
      corpo: zod.string(),
      autore: zod.string(),
      autoreClerkId: zod.string().nullish(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Elimina discussione
 */
export const DeleteForumThreadParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Aggiungi risposta a una discussione
 */
export const CreateForumReplyParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateForumReplyBody = zod.object({
  corpo: zod.string(),
  autore: zod.string(),
  autoreClerkId: zod.string().nullish(),
});

/**
 * @summary Elimina risposta
 */
export const DeleteForumReplyParams = zod.object({
  threadId: zod.coerce.number(),
  replyId: zod.coerce.number(),
});