# lib/api-zod/src/generated/types/createForumThreadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for sito annunci
 * OpenAPI spec version: 0.1.0
 */

export interface CreateForumThreadBody {
  titolo: string;
  corpo: string;
  categoria: string;
  autore: string;
  /** @nullable */
  autoreClerkId?: string | null;
}