# lib/api-zod/src/generated/types/forumListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for sito annunci
 * OpenAPI spec version: 0.1.0
 */
import type { ForumThread } from "./forumThread";

export interface ForumListResponse {
  threads: ForumThread[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}