# lib/api-zod/src/generated/types/forumReply.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for sito annunci
 * OpenAPI spec version: 0.1.0
 */

export interface ForumReply {
  id: number;
  threadId: number;
  corpo: string;
  autore: string;
  /** @nullable */
  autoreClerkId?: string | null;
  createdAt: Date;
}