# lib/api-zod/src/generated/types/forumThreadWithReplies.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for sito annunci
 * OpenAPI spec version: 0.1.0
 */
import type { ForumReply } from "./forumReply";
import type { ForumThread } from "./forumThread";

export interface ForumThreadWithReplies {
  thread: ForumThread;
  replies: ForumReply[];
}